requirements.txt

anyio==4.9.0
aiohttp
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
eventlet==0.40.0
exceptiongroup==1.3.0
Flask==3.1.1
Flask-SocketIO==5.5.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.24.4
pandas==2.0.3
pandas_ta==0.3.14b0
python-dateutil==2.9.0.post0
python-engineio==4.12.1
python-socketio==5.13.0
python-telegram-bot==20.7
pytz==2025.2
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
ta==0.10.2
tabulate
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
websockets==10.3
Werkzeug==3.1.3
wsproto==1.2.0
